Added JWT claims

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import get_jwt_identity, jwt_required, get_jwt
 from models.item import ItemModel
 
 
@@ -17,8 +17,8 @@ class Item(Resource):
                         )
 
     @jwt_required()
-    def get(self, name: str):
-        item = ItemModel.find_by_name(name=name)
+    def get(self, name):
+        item = ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {'message': 'Item not found'}, 404
@@ -38,13 +38,17 @@ class Item(Resource):
 
         return item.json()
 
+    @jwt_required()
     def delete(self, name: str):
-        item = ItemModel.find_by_name(name=name)
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
 
+        item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
-
-        return {'message': 'Item deleted'}
+            return {'message': 'Item deleted.'}
+        return {'message': 'Item not found.'}, 404
 
     def put(self, name: str):
         data = Item.parser.parse_args()
